Funcionalidades da atividades concluídas.

+++ src/main/java/br/com/ifma/imobiliaria/execute/LocacaoExecute.java
@@ -6,6 +6,7 @@ import br.com.ifma.imobiliaria.model.Locacao;
 import br.com.ifma.imobiliaria.service.ClienteService;
 import br.com.ifma.imobiliaria.service.ImovelService;
 import br.com.ifma.imobiliaria.service.LocacaoService;
+import br.com.ifma.imobiliaria.util.DataConvert;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -38,6 +39,11 @@ public class LocacaoExecute {
         locacao.setAtivo(true);
         locacao.setObservacoes("Imovel está alugado");
         locacao.setDiaVencimento(10);
+        locacao.setPercentualMulta(new BigDecimal(5));
+        locacao.setDiaVencimento(10);
+        locacao.setDataInicio(new DataConvert().converterData("15/05/2020"));
+        locacao.setDataFim(new DataConvert().converterData("15/04/2021"));
+        locacao.setValorAluguel(new BigDecimal(1000));
 
         locacao1 = locacaoService.listarLocacao(cliente);
 
@@ -46,16 +52,16 @@ public class LocacaoExecute {
             //loc.setPercentualMulta(new BigDecimal(15.5));
             //locacaoService.save(loc);
 
-            System.out.println(locacaoService.imprimiLocacao(loc));
+            //System.out.println(locacaoService.imprimiLocacao(loc));
             System.out.println();
 
-            /*if (loc.getId().equals(1)){
+            if (loc.getId().equals(3)){
                 locacaoService.remove(loc.getId());
-            }*/
+            }
 
         }
 
-        //locacaoService.save(locacao);
+        locacaoService.save(locacao);
     }
